Another preliminary task: Moved interfaces IEntity and IRPCEntity into package dk.ule.oapenwb.entity for easier outsourcing into the new oapenwb-persistency project
