+++ oapenwb-javalin/src/main/java/dk/ule/oapenwb/entity/IRPCEntity.java
@@ -1,6 +1,6 @@
 // SPDX-FileCopyrightText: © 2022 Michael Köther <mkoether38@gmail.com>
 // SPDX-License-Identifier: AGPL-3.0-only
-package dk.ule.oapenwb.logic.admin.lexeme;
+package dk.ule.oapenwb.entity;
 
 import com.fasterxml.jackson.annotation.JsonView;
 import dk.ule.oapenwb.entity.Views;
