+++ oapenwb-javalin/src/main/java/dk/ule/oapenwb/entity/content/basedata/LinkType.java
@@ -7,7 +7,7 @@ import com.fasterxml.jackson.annotation.JsonView;
 import dk.ule.oapenwb.entity.Views;
 import dk.ule.oapenwb.entity.content.lexemes.Link;
 import dk.ule.oapenwb.entity.content.lexemes.SynLink;
-import dk.ule.oapenwb.logic.admin.generic.IEntity;
+import dk.ule.oapenwb.entity.IEntity;
 import io.hypersistence.utils.hibernate.type.json.JsonBinaryType;
 import jakarta.persistence.*;
 import lombok.AllArgsConstructor;
