+++ oapenwb-javalin/src/main/java/dk/ule/oapenwb/entity/content/lexemes/SynGroup.java
@@ -6,8 +6,8 @@ import com.fasterxml.jackson.annotation.JsonIgnore;
 import com.fasterxml.jackson.annotation.JsonView;
 import dk.ule.oapenwb.entity.Views;
 import dk.ule.oapenwb.entity.ApiAction;
-import dk.ule.oapenwb.logic.admin.generic.IEntity;
-import dk.ule.oapenwb.logic.admin.lexeme.IRPCEntity;
+import dk.ule.oapenwb.entity.IEntity;
+import dk.ule.oapenwb.entity.IRPCEntity;
 import io.hypersistence.utils.hibernate.type.json.JsonBinaryType;
 import jakarta.persistence.*;
 import lombok.Data;
