+++ oapenwb-javalin/src/main/java/dk/ule/oapenwb/entity/ui/UiResultCategory.java
@@ -5,7 +5,7 @@ package dk.ule.oapenwb.entity.ui;
 import com.fasterxml.jackson.annotation.JsonIgnore;
 import com.fasterxml.jackson.annotation.JsonView;
 import dk.ule.oapenwb.entity.Views;
-import dk.ule.oapenwb.logic.admin.generic.IEntity;
+import dk.ule.oapenwb.entity.IEntity;
 import jakarta.persistence.*;
 import lombok.AllArgsConstructor;
 import lombok.Data;
