+++ oapenwb-javalin/src/main/java/dk/ule/oapenwb/logic/admin/common/IRestController.java
@@ -3,7 +3,7 @@
 package dk.ule.oapenwb.logic.admin.common;
 
 import dk.ule.oapenwb.base.error.CodeException;
-import dk.ule.oapenwb.logic.admin.generic.IEntity;
+import dk.ule.oapenwb.entity.IEntity;
 import dk.ule.oapenwb.logic.context.Context;
 
 import java.util.List;
