+++ oapenwb-javalin/src/main/java/dk/ule/oapenwb/logic/admin/generic/CEntityController.java
@@ -4,6 +4,7 @@ package dk.ule.oapenwb.logic.admin.generic;
 
 import com.google.inject.Singleton;
 import dk.ule.oapenwb.base.error.CodeException;
+import dk.ule.oapenwb.entity.IEntity;
 import dk.ule.oapenwb.logic.context.Context;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
