+++ oapenwb-javalin/src/main/java/dk/ule/oapenwb/logic/admin/generic/ICEntityController.java
@@ -2,6 +2,8 @@
 // SPDX-License-Identifier: AGPL-3.0-only
 package dk.ule.oapenwb.logic.admin.generic;
 
+import dk.ule.oapenwb.entity.IEntity;
+
 import java.io.Serializable;
 
 public interface ICEntityController<T extends IEntity<S>, S extends Serializable> extends IEntityController<T, S>
