+++ oapenwb-javalin/src/main/java/dk/ule/oapenwb/logic/admin/generic/IEntityController.java
@@ -2,6 +2,7 @@
 // SPDX-License-Identifier: AGPL-3.0-only
 package dk.ule.oapenwb.logic.admin.generic;
 
+import dk.ule.oapenwb.entity.IEntity;
 import dk.ule.oapenwb.logic.admin.common.IRestController;
 
 public interface IEntityController<T extends IEntity<S>, S> extends IRestController<T, S>
