+++ oapenwb-javalin/src/main/java/dk/ule/oapenwb/logic/admin/generic/IGroupedEntitySupplier.java
@@ -3,6 +3,7 @@
 package dk.ule.oapenwb.logic.admin.generic;
 
 import dk.ule.oapenwb.base.error.CodeException;
+import dk.ule.oapenwb.entity.IEntity;
 
 import java.io.Serializable;
 import java.util.List;
